% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\title{Estimate spectrum...}
\usage{
spectrum(x, y, config)
}
\arguments{
\item{x}{Data frame with...}

\item{y}{Data frame...}

\item{config}{Configuration...}
}
\value{
named list...
}
\description{
This function ...
}
